backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0,<2.3.0
opencv-python-headless>=4.8.0
sqlalchemy>=2.0.0
alembic>=1.10.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.25.0
pymysql>=1.0.0
cryptography>=3.4.0